+++ back-end/src/database/migrations/20220920201233-create-products.js
@@ -14,7 +14,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(4, 2)
       },
       url_image: {
         allowNull: false,
